Moved fullOuterJoinDataFrames method from SparkCompare.scala to DiffResult.scala
