+++ src/main/scala/org/finra/msd/sparkcompare/SparkCompare.scala
@@ -148,7 +148,7 @@ object SparkCompare {
     val inLnotinR = groupedLeft.except(groupedRight).toDF()
     val inRnotinL = groupedRight.except(groupedLeft).toDF()
 
-    return new DiffResult(inLnotinR, inRnotinL)
+    return DiffResult(inLnotinR, inRnotinL)
   }
 
   /**
@@ -166,41 +166,5 @@ object SparkCompare {
     val countsPair = new CountResult(leftCount, rightCount)
 
     return countsPair
-  }
-
-  /**
-    * This method does a full outer join between the resulting left and right dataframes from the method
-    * SparkCompare.compareSchemaDataFrames. It will return a single dataframe having the left columns prefixed with l_
-    * and the right columns prefixed with r_. the Key columns will not have prefixed. The resulting dataframe will have
-    * all l_ columns on the left, then the Key columns in the middle, then the r_ columns on the right.
-    *
-    * @param left             Dataframe having inLeftNotInRight resulting from SparkCompare.compareSchemaDataFrames
-    * @param right            Dataframe having inRightNotInLeft resulting from SparkCompare.compareSchemaDataFrames
-    * @param compositeKeyStrs a Sequence of Strings having the primary keys applicable for both dataframes
-    * @return a dataframe having the resulting full outer join operation.
-    */
-  def fullOuterJoinDataFrames(left: DataFrame, right: DataFrame, compositeKeyStrs: Seq[String]): DataFrame = {
-
-    //convert column names to uppercase
-    val upperCaseLeft: DataFrame = left.toDF(left.columns.map(_.toUpperCase): _*)
-    val upperCaseRight: DataFrame = right.toDF(right.columns.map(_.toUpperCase): _*)
-
-    val compositeKeysUpperCase: Seq[String] = compositeKeyStrs.map(k => k.toUpperCase)
-    val nonKeyCols: Seq[String] = upperCaseLeft.columns.filter(c => !compositeKeysUpperCase.contains(c)).toSeq
-
-    //prepend l and r to nonkey columns
-    val prependedColumnsLeft = compositeKeysUpperCase ++ nonKeyCols.map(c => "l_" + c).toSeq
-    val prependedColumnsRight = compositeKeysUpperCase ++ nonKeyCols.map(c => "r_" + c).toSeq
-
-    //reselect the dataframes with prepended l. & r. to the columnss
-    val prependedLeftDf: DataFrame = upperCaseLeft.toDF(prependedColumnsLeft: _*)
-    val prependedRightDf: DataFrame = upperCaseRight.toDF(prependedColumnsRight: _*)
-
-
-    val joinedDf: DataFrame = prependedLeftDf.as("l")
-      .join(prependedRightDf.as("r"), compositeKeysUpperCase, "full_outer")
-
-    val allColsWithKeysInTheMiddle: Seq[String] = nonKeyCols.map(c => "l_" + c) ++ compositeKeysUpperCase ++ nonKeyCols.map(c => "r_" + c)
-    joinedDf.select(allColsWithKeysInTheMiddle.map(name => col(name)): _*)
   }
 }
